OCSADV-410: Changes as per code review.
