+++ bundle/edu.gemini.wdba.xmlrpc.server/src/main/java/edu/gemini/wdba/tcc/TargetGroupConfig.java
@@ -5,6 +5,7 @@
 package edu.gemini.wdba.tcc;
 
 import edu.gemini.shared.util.immutable.ImList;
+import edu.gemini.shared.util.immutable.ImOption;
 import edu.gemini.shared.util.immutable.Option;
 import edu.gemini.spModel.guide.GuideProbe;
 import edu.gemini.spModel.target.SPTarget;
@@ -13,6 +14,7 @@ import edu.gemini.spModel.target.env.TargetEnvironment;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * The {@link ParamSet} implementation for a group of targets.
@@ -22,38 +24,26 @@ public final class TargetGroupConfig extends ParamSet {
 
     public static TargetGroupConfig createBaseGroup(final TargetEnvironment env) {
         final ImList<SPTarget> targets = env.getUserTargets().cons(env.getBase());
-        return new TargetGroupConfig(TccNames.BASE, targets, GuideProbeTargets.NO_TARGET,
-                env.getBase().getTarget().getName());
+        return new TargetGroupConfig(TccNames.BASE, targets, ImOption.apply(env.getBase()));
     }
 
     public static TargetGroupConfig createGuideGroup(final GuideProbeTargets gt) {
-        final GuideProbe guider = gt.getGuider();
-        final ImList<SPTarget> manualTargets = gt.getManualTargets();
+        final String tag = TargetConfig.getTag(gt.getGuider());
+        final ImList<SPTarget> targets = gt.getTargets();
         final Option<SPTarget> primaryOpt = gt.getPrimary();
-        final Option<SPTarget> bagsTargetOpt = gt.getBagsTarget();
-
-        // SW: no longer always setting a primary target.
-//        if ((primary == null) && (targets.size() > 0)) primary = targets.head();
-
-        final String primaryTargetName = primaryOpt.map(p -> p.getTarget().getName()).getOrNull();
-        final String tag = TargetConfig.getTag(guider);
-        return new TargetGroupConfig(tag, manualTargets, bagsTargetOpt, primaryTargetName);
+        return new TargetGroupConfig(tag, targets, primaryOpt);
     }
 
-    private TargetGroupConfig(final String name, final ImList<SPTarget> targets,
-                              final Option<SPTarget> bagsTarget, final String primaryTargetName) {
+    private TargetGroupConfig(final String name, final ImList<SPTarget> targets, final Option<SPTarget> primaryTarget) {
         super(name);
 
         addAttribute(TYPE, TYPE_VALUE);
 
-        if ((primaryTargetName != null) && !"".equals(primaryTargetName)) {
-            putParameter(TccNames.PRIMARY, primaryTargetName);
-        }
+        primaryTarget.map(t -> t.getTarget().getName())
+                .filter(n -> !"".equals(n))
+                .foreach(n -> putParameter(TccNames.PRIMARY, n));
 
-        final List<String> targetNames = new ArrayList<String>(targets.size());
-        targets.foreach(t -> targetNames.add(t.getTarget().getName()));
+        final List<String> targetNames = targets.toList().stream().map(t -> t.getTarget().getName()).collect(Collectors.toList());
         putParameter(TccNames.TARGETS, targetNames);
-
-        bagsTarget.foreach(t -> putParameter(TccNames.BAGSTARGET, t.getTarget().getName()));
     }
 }
