+++ bundle/edu.gemini.wdba.xmlrpc.server/src/main/java/edu/gemini/wdba/tcc/TccNames.java
@@ -23,7 +23,6 @@ public final class TccNames {
     public static final String AZEL = "AzEl";
     public static final String BASE = "Base";
 //    public static final String BASE = "sciencetarget";
-    public static final String BAGSTARGET = "bagsTarget";
     public static final String BASICCHOP = "BasicChop";
     public static final String BRIGHTNESS = "brightness";
     //public static final String FIELD = "targetEnv";
