+++ bundle/jsky.app.ot/src/main/java/jsky/app/ot/tpe/GemsGuideStarWorker.java
@@ -63,16 +63,6 @@ public class GemsGuideStarWorker extends SwingWorker implements MascotProgress {
     }
 
     /**
-     * Create an instance for one time use.
-     */
-    public GemsGuideStarWorker() {
-        ProgressPanel progressPanel = ProgressPanel.makeProgressPanel("GeMS Strehl Calculations",
-                TpeManager.create().getImageWidget());
-        progressPanel.addActionListener(e -> interrupted = true);
-        init(progressPanel);
-    }
-
-    /**
      * Called by constructors
      */
     public void init(StatusLogger statusLogger) {
